+++ components/DefaultLayout.tsx
@@ -2,7 +2,7 @@ import Header from './Header';
 import Footer from './Footer';
 import { PropsWithChildren } from 'react';
 
-export default function Layout({ children }: PropsWithChildren) {
+export default function DefaultLayout({ children }: PropsWithChildren) {
   return (
     <>
       <Header />