+++ pages/_app.tsx
@@ -1,10 +1,15 @@
 import type { AppProps } from "next/app";
 import Head from "next/head";
 import "../styles/style.css";
-import Layout from "../components/Layout";
+import DefaultLayout from "../components/DefaultLayout";
 import { appWithTranslation } from "next-i18next";
+import NoHeaderLayout from "../components/NoHeaderLayout";
 
-function App({ Component, pageProps }: AppProps) {
+function App({ Component, pageProps, router }: AppProps) {
+  const Layout = router.pathname.includes("/invite")
+    ? NoHeaderLayout
+    : DefaultLayout;
+
   return (
     <Layout>
       <Head>